cherry picked NivekT's move of test case to test_local_io.py
