+++ test/_fakedata/_create_fake_data.py
@@ -4,7 +4,6 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-import io
 import os
 import tarfile
