+++ test/test_datapipe.py
@@ -41,7 +41,6 @@ from torchdata.datapipes.iter import (
     Rows2Columnar,
     SampleMultiplexer,
     UnZipper,
-    WebDataset,
 )
 from torchdata.datapipes.map import MapDataPipe
 
@@ -876,66 +875,6 @@ class TestDataPipe(expecttest.TestCase):
             dup_map_dp._load_map()
             self.assertEqual(len(wa), 1)
             self.assertRegex(str(wa[0].message), r"Found duplicate key")
-
-    def test_webdataset(self) -> None:
-        # Functional Test: groups samples correctly
-        source_dp = IterableWrapper(
-            # simulated tar file content
-            [
-                ("/path/to/file1.jpg", b"1"),
-                ("/path/to/_something_", b"nothing"),
-                ("/path/to/file1.cls", b"2"),
-                ("/path/to/file2.jpg", b"3"),
-                ("/path/to/file2.cls", b"4"),
-            ]
-        )
-        web_dataset = WebDataset(source_dp)
-        self.assertEqual(
-            # expected grouped output
-            [
-                {".jpg": b"1", ".cls": b"2", "__key__": "/path/to/file1"},
-                {".jpg": b"3", ".cls": b"4", "__key__": "/path/to/file2"},
-            ],
-            list(web_dataset),
-        )
-
-    def test_webdataset2(self) -> None:
-        tmpdir = str(self.tmpdir)
-
-        def add_data_to_tar(archive, name, value):
-            if isinstance(value, str):
-                value = value.encode()
-            info = tarfile.TarInfo(name)
-            info.size = len(value)
-            archive.addfile(info, io.BytesIO(value))
-
-        def create_wds_tar(dest):
-            with tarfile.open(dest, mode="w") as archive:
-                for i in range(10):
-                    add_data_to_tar(archive, f"data/{i}.txt", f"text{i}")
-                    add_data_to_tar(archive, f"data/{i}.bin", f"bin{i}")
-
-        dest = os.path.join(tmpdir, "wds.tar")
-        create_wds_tar(dest)
-
-        def decode(item):
-            key, value = item
-            if key.endswith(".txt"):
-                return key, value.read().decode("utf-8")
-            if key.endswith(".bin"):
-                return key, value.read().decode("utf-8")
-
-        datapipe1 = FileLister(tmpdir, "wds*.tar")
-        datapipe2 = FileOpener(datapipe1, mode="b")
-        dataset = datapipe2.load_from_tar().map(decode).webdataset()
-        items = list(dataset)
-        assert len(items) == 10
-        assert items[0][".txt"] == "text0"
-        assert items[9][".bin"] == "bin9"
-
-    @pytest.fixture(autouse=True)
-    def initdir(self, tmpdir):
-        self.tmpdir = tmpdir
 
 
 if __name__ == "__main__":
